requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.59.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
matplotlib==3.10.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
